+++ NEWS.md
@@ -1,6 +1,6 @@
 # gsignal 0.3-2
 
-- date: 20210509
+- date: 20210515
 - corrected CRAN WARNINGs on ATLAS, MKL, valgrind, fedora, solaris
 - corrected import NOTE for grDevices
 - adapted code in vignette "gsignal"
@@ -10,6 +10,8 @@
 - use matrix() instead of as.matrix() in functions
     dct(), idct(), czt(), dst(), idst(), fht(), ifht()
 - minor bugfix in arburg()
+- bugfix in filter.cpp: resize a and b vectors length of zi
+
 
 ---